Use size_t for matrix sizes.
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl@7545 7226f50c-d813-0410-99bf-979b288ab488

+++ main/Ravl/Core/I386SSE/VectorUtilsI386SSE.cc
@@ -480,9 +480,9 @@ namespace RavlBaseVectorN {
 
   void SSEConvolveKernelF(const float *vi, // Scanned image, probably not aligned.
                           const float *vk, // Kernel, expected to be aligned.
-                          unsigned int rows,
-                          unsigned int cols,
-                          unsigned int byteStride,
+                          size_t rows,
+                          size_t cols,
+                          int byteStride,
                           float *result
                          )
   {
